Revert "consolidate tests"
This reverts commit b7b6ce61b2591a0f1a6a96dca38cc92686ec343b.
