+++ scripts/which_dynamo.py
@@ -1,0 +1,3 @@
+import importlib
+mod = importlib.import_module('API.dynamo')
+print('module file=', getattr(mod,'__file__',None))
