+++ tests/test_api_dynamo_more_errors.py
@@ -1,0 +1,2 @@
+# moved to tests/test_errors_consolidated.py
+
