+++ tests/test_artifact_downloader_errors.py
@@ -1,0 +1,2 @@
+# file consolidated into tests/test_errors_consolidated.py
+
