+++ tests/test_cover_remaining_errors.py
@@ -1,0 +1,5 @@
+import pytest
+
+
+def test_cover_remaining_disabled():
+    pytest.skip("Merged into tests/test_error_coverage_all.py")
