+++ tests/test_force_mark_remaining.py
@@ -1,0 +1,5 @@
+import pytest
+
+
+def test_force_mark_remaining_disabled():
+    pytest.skip("Mechanical force-marking disabled — using real tests only")
