+++ tests/test_registry_handler_errors.py
@@ -1,0 +1,2 @@
+# moved to tests/test_errors_consolidated.py
+
