+++ tests/test_server.py
@@ -10,11 +10,7 @@ from unittest.mock import Mock, patch, MagicMock
 import requests
 import server
 server.requests = requests
-from fastapi import HTTPException
-from fastapi.testclient import TestClient
-import pytest
-import asyncio
-from handlers import registry_handler
+
 
 # Add paths
 SCRIPT_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
@@ -1054,317 +1050,6 @@ class TestDeleteArtifactEndpoint(unittest.TestCase):
         mock_delete.assert_called_once_with(name)
         # Ensure it gets tracked in DELETED_ARTIFACTS
         self.assertIn(name, server.DELETED_ARTIFACTS)
-
-def test_size_calc_github_api_and_unexpected(monkeypatch, caplog):
-    caplog.set_level("ERROR")
-    # Make requests.get raise RequestException to hit GitHub API error path
-    from requests.exceptions import RequestException
-
-    def raise_req(*a, **k):
-        raise RequestException("netfail")
-
-    monkeypatch.setattr("requests.get", raise_req)
-    # Call size calc; should raise or return but log the GitHub API error
-    try:
-        server._calculate_artifact_size_api("https://github.com/owner/repo", "model")
-    except Exception:
-        pass
-    assert any("GitHub API error" in rec.message or "GitHub API" in rec.message for rec in caplog.records)
-
-    # Now make generic exception (still inside GitHub handling -> logs same GitHub API error)
-    def raise_gen(*a, **k):
-        raise ValueError("boom")
-
-    monkeypatch.setattr("requests.get", raise_gen)
-    try:
-        server._calculate_artifact_size_api("https://github.com/owner/repo", "model")
-    except Exception:
-        pass
-    assert any("GitHub API error" in rec.message or "GitHub API" in rec.message for rec in caplog.records)
-
-
-def test_artifact_by_regex_search_error(monkeypatch, caplog):
-    caplog.set_level("ERROR")
-
-    # Fake request with JSON body
-    class Req:
-        def __init__(self, body):
-            self._body = body
-            self.headers = {}
-
-        async def json(self):
-            return self._body
-
-    # Make re.compile return an object whose search raises
-    def fake_compile(regex, flags=0):
-        class P:
-            def search(self, s):
-                raise RuntimeError("search fail")
-
-        return P()
-
-    monkeypatch.setattr(server, "re", server.re)
-    monkeypatch.setattr(server, "re", server.re)
-    monkeypatch.setattr("re.compile", fake_compile)
-
-    # Ensure registry returns one artifact to iterate
-    monkeypatch.setattr("handlers.registry_handler.list_artifacts", lambda *a, **k: [{"name": "a"}])
-
-    req = Req({"regex": "a"})
-    with pytest.raises(HTTPException):
-        # call the async endpoint
-        import asyncio
-
-        asyncio.run(server.artifact_by_regex(req))
-
-    assert any("Regex search error" in rec.message or "Regex search timed out" in rec.message for rec in caplog.records)
-
-
-def test_rate_request_invalid_id_logs(monkeypatch, caplog):
-    caplog.set_level("ERROR")
-    # Ensure list_artifacts returns empty so lookup fails
-    monkeypatch.setattr("handlers.registry_handler.list_artifacts", lambda *a, **k: [])
-    import asyncio
-
-    with pytest.raises(HTTPException):
-        asyncio.run(server.rate_model("notanint", Mock()))
-
-    assert any("Invalid ID format" in rec.message or "Artifact 'notanint' not found" in rec.message for rec in caplog.records)
-
-
-
-
-def test_license_check_errors(monkeypatch, caplog):
-    caplog.set_level("ERROR")
-
-    def bad_get(*a, **k):
-        raise server.requests.RequestException("boom")
-
-    # Ensure artifact exists check passes
-    monkeypatch.setattr(server, "_verify_artifact_exists", lambda *_a, **_k: True)
-
-    monkeypatch.setattr(server, "requests", server.requests)
-    monkeypatch.setattr(server.requests, "get", bad_get)
-
-    r = client.post("/artifact/model/1/license-check", json={"github_url": "https://github.com/owner/repo"})
-    assert r.status_code == 502
-    assert any("Error fetching GitHub page" in r.message for r in caplog.records if r.levelname == "ERROR")
-
-
-def test_calculate_size_unexpected_error(monkeypatch, caplog):
-    caplog.set_level("ERROR")
-
-    def bad_get(*a, **k):
-        raise Exception("boom")
-
-    monkeypatch.setattr(server, "requests", server.requests)
-    monkeypatch.setattr(server.requests, "get", bad_get)
-
-    # Use a GitHub URL to exercise the GitHub branch which will now raise
-    res = server._calculate_artifact_size_api("https://github.com/owner/repo", "model")
-    assert res == 0.0
-    assert any("GitHub API error" in r.message or "Unexpected error for" in r.message for r in caplog.records if r.levelname == "ERROR")
-
-
-def test_log_audit_event_failure(monkeypatch, caplog):
-    caplog.set_level("ERROR")
-
-    monkeypatch.setattr(server.registry_handler, "get_artifact_by_id", lambda *_a, **_k: (_ for _ in ()).throw(Exception("boom")))
-    server._log_audit_event(1, "x", "model", "AUDIT")
-    assert any("Failed to log audit event" in r.message for r in caplog.records if r.levelname == "ERROR")
-
-
-def test_lineage_artifact_not_found_and_invalid(monkeypatch, caplog):
-    caplog.set_level("ERROR")
-
-    # Not found (numeric but not present)
-    monkeypatch.setattr(server.registry_handler, "list_artifacts", lambda: [])
-    r = client.get("/artifact/model/123/lineage")
-    assert r.status_code == 404
-    assert any("Lineage request - Artifact not found" in r.message for r in caplog.records if r.levelname == "ERROR")
-
-    caplog.clear()
-    # Invalid format
-    r = client.get("/artifact/model/notanumber/lineage")
-    assert r.status_code == 404
-    assert any("Lineage request - Invalid ID format" in r.message for r in caplog.records if r.levelname == "ERROR")
-
-def test_rate_model_error_listing(monkeypatch, caplog):
-    caplog.set_level("ERROR")
-
-    def boom():
-        raise Exception("boom")
-
-    monkeypatch.setattr(server.registry_handler, "list_artifacts", boom)
-
-    r = client.get("/artifact/model/1/rate")
-    assert r.status_code == 500
-    assert any("Error listing artifacts: boom" in r.message for r in caplog.records if r.levelname == "ERROR")
-
-
-def test_rate_model_invalid_id_format(monkeypatch, caplog):
-    caplog.set_level("ERROR")
-
-    monkeypatch.setattr(server.registry_handler, "list_artifacts", lambda: [])
-
-    r = client.get("/artifact/model/notanumber/rate")
-    assert r.status_code == 404
-    assert any("Invalid ID format" in r.message for r in caplog.records if r.levelname == "ERROR")
-
-
-def test_rate_model_not_found_logs_errors(monkeypatch, caplog):
-    caplog.set_level("ERROR")
-
-    # Return empty list so numeric lookup fails
-    monkeypatch.setattr(server.registry_handler, "list_artifacts", lambda: [])
-
-    r = client.get("/artifact/model/9999/rate")
-    assert r.status_code == 404
-    # Should log artifact not found messages
-    assert any("Artifact '" in r.message for r in caplog.records if r.levelname == "ERROR")
-
-
-
-
-def test_startup_registry_init_failure(monkeypatch, caplog):
-    # init_registry raises -> startup_event should log and re-raise
-    monkeypatch.setattr("handlers.registry_handler.init_registry", lambda: (_ for _ in ()).throw(Exception("init fail")))
-    caplog.clear()
-    with pytest.raises(Exception):
-        asyncio.run(server.startup_event())
-    assert any("Failed to initialize registry" in rec.message for rec in caplog.records)
-
-
-def test_rate_model_group_urls_error(monkeypatch, caplog):
-    # group_urls_by_type raises -> should be caught and logged
-    monkeypatch.setattr(server.model_evaluator.url_classifier, "group_urls_by_type", lambda *a, **k: (_ for _ in ()).throw(Exception("group fail")))
-    monkeypatch.setattr(server.model_evaluator, "evaluate_urls", lambda *a, **k: [])
-    monkeypatch.setattr("handlers.registry_handler.get_artifact_by_id", lambda *a, **k: None)
-    caplog.clear()
-    asyncio.run(server.rate_model_background(1, "n", "u"))
-    assert any("URL classification error" in rec.message or "URL classification" in rec.message for rec in caplog.records)
-
-
-def test_rate_model_evaluate_raises(monkeypatch, caplog):
-    # evaluate_urls raises -> outer except should log Error rating model
-    monkeypatch.setattr(server.model_evaluator.url_classifier, "group_urls_by_type", lambda *a, **k: {})
-    monkeypatch.setattr(server.model_evaluator, "evaluate_urls", lambda *a, **k: (_ for _ in ()).throw(Exception("eval fail")))
-    monkeypatch.setattr("handlers.registry_handler.get_artifact_by_id", lambda *a, **k: None)
-    caplog.clear()
-    asyncio.run(server.rate_model_background(2, "nm", "u"))
-    assert any("Error rating model" in rec.message for rec in caplog.records)
-
-
-def test_check_license_fetch_and_general_errors(monkeypatch, caplog):
-    # RequestException path
-    from requests.exceptions import RequestException
-
-    def raise_req(*a, **k):
-        raise RequestException("net")
-
-    monkeypatch.setattr("requests.get", raise_req)
-    caplog.clear()
-    with pytest.raises(HTTPException):
-        server._check_license_compatibility("https://github.com/owner/repo")
-    assert any("Error fetching GitHub page" in rec.message for rec in caplog.records)
-
-    # General exception path
-    def raise_val(*a, **k):
-        raise ValueError("boom")
-
-    monkeypatch.setattr("requests.get", raise_val)
-    caplog.clear()
-    with pytest.raises(HTTPException):
-        server._check_license_compatibility("https://github.com/owner/repo")
-    assert any("Error checking license" in rec.message for rec in caplog.records)
-
-
-def test_lineage_invalid_id_and_json_decode(monkeypatch, caplog):
-    # Invalid ID format
-    req = Mock(); req.headers = {}
-    caplog.clear()
-    with pytest.raises(HTTPException):
-        server.get_artifact_lineage("notanint", req)
-    assert any("Lineage request - Invalid ID format" in rec.message for rec in caplog.records)
-
-    # JSON decode error path deeper in lineage
-    # Prepare a matching artifact whose metadata_json is malformed
-    bad = {"name": "x", "metadata_json": "{bad json"}
-    monkeypatch.setattr("handlers.registry_handler.list_artifacts", lambda *a, **k: [bad])
-    req2 = Mock(); req2.headers = {}
-    caplog.clear()
-    with pytest.raises(HTTPException):
-        server.get_artifact_lineage(str(server.gen_id("x")), req2)
-    assert any("JSON decode error" in rec.message or "Unexpected error for artifact" in rec.message for rec in caplog.records)
-
-
-def test_log_audit_event_failure(monkeypatch, caplog):
-    # Cause json.loads to raise by returning malformed metadata_json
-    bad = {"name": "a", "metadata_json": "{bad"}
-    monkeypatch.setattr("handlers.registry_handler.get_artifact_by_id", lambda *a, **k: bad)
-    caplog.clear()
-    server._log_audit_event(1, "a", "model", "CREATE")
-    assert any("Failed to log audit event" in rec.message for rec in caplog.records)
-
-
-def test_startup_registry_init_failure(monkeypatch):
-    def fail():
-        raise RuntimeError("init fail")
-
-    monkeypatch.setattr(registry_handler, "init_registry", fail)
-
-    with pytest.raises(RuntimeError):
-        # call the startup event directly
-        asyncio.run(server.startup_event())
-
-
-def test_rate_model_group_urls_error(caplog, monkeypatch):
-    caplog.set_level("ERROR")
-
-    # make url_classifier.group_urls_by_type raise
-    class Bad:
-        def group_urls_by_type(self, urls):
-            raise Exception("boom")
-
-    monkeypatch.setattr(server.model_evaluator, "url_classifier", Bad())
-
-    # run the async background rating (just execute the path)
-    asyncio.run(server.rate_model_background(12345, "name", "https://huggingface.co/model"))
-
-
-def test_rate_model_evaluate_urls_error(caplog, monkeypatch):
-    caplog.set_level("ERROR")
-
-    def bad_eval(urls, artifact_id=None):
-        raise Exception("eval boom")
-
-    monkeypatch.setattr(server.model_evaluator, "evaluate_urls", bad_eval)
-
-    asyncio.run(server.rate_model_background(123, "name", "https://huggingface.co/model"))
-
-
-
-def test_artifacts_bad_json_logs_error(caplog):
-    caplog.clear()
-    # send invalid json bytes
-    resp = client.post("/artifacts", data=b"{badjson", headers={"Content-Type": "application/json"})
-    assert resp.status_code == 400
-    assert any("JSON parse error" in r.message or "Unexpected error" in r.message for r in caplog.records)
-
-
-def test_lineage_invalid_id_logs_error(caplog):
-    caplog.clear()
-    resp = client.get("/artifact/model/notanint/lineage")
-    assert resp.status_code == 404
-    assert any("Lineage request - Invalid ID format" in r.message for r in caplog.records)
-
-
-def test_rate_request_invalid_id_logs_error(caplog):
-    caplog.clear()
-    # call rate endpoint with invalid id via post to trigger earlier rate request error path
-    resp = client.post("/artifact/model/invalid-id/rate", json={})
-    # endpoint may not exist; accept 404 or 400
-    assert resp.status_code in (400, 404, 405)
 
 
 if __name__ == "__main__":
